docs: correct completeness claims; move last inline test block
Review findings 1-4 on this PR.

1. The rule-9 node claimed #657 left no inline test blocks and every test
   file #[path]-wired. Both false: src/messages.rs had an inline
   `mod from_str_tests`, and ~22 `mod tests;` declarations still resolve to
   <dir>/tests.rs. Moved the inline block to src/messages/tests.rs and
   restated both this node and testing/sibling-test-files.md, which carried
   the same "fully applied" overstatement from the testing/ pass.
2. docs/build-and-test.md said CI's doc leg "builds without warnings" — it
   runs without RUSTDOCFLAGS, which is exactly what the new node corrects.
3. Added `cargo build --examples` (both configs) to the gate; `just test`
   never compiles examples, and they are a caller surface.
4. Reframed the node's Why around which half of the list CI re-runs and
   which half has no safety net.

Verification: just test green on all three legs (1367/1384/1713 unit +
203/208/307 integration-compile), clippy default + sync clean, rules-check
22 nodes.
